Fix locales on Ubuntu 14.04
The unittests use pt_BR and pt_BR.UTF8 but Ubuntu14 (where the travis
tests run) do not know about those locales.  Install the Portuguese
language pack for UTF8 support and use locale-gen pt_BR to generate the
latin-1 version of the locale.  This allows the unittests to run and
give predictable output.